+++ Sources/FetchNodeDetails/FetchNodeDetails+promise.swift
