+++ Sources/FetchNodeDetails/Statically Typed/EthereumNetwork.swift
